+++ previous-versions/files/src/sdk/models/shared/schema.ts
@@ -5,10 +5,7 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { Expose } from "class-transformer";
 
-/**
- * The request made is not valid.
- */
-export class ErrorMessage extends SpeakeasyBase {
+export class Schema extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "canBeRetried" })
     canBeRetried?: string;